% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastxml-package.R
\docType{package}
\name{blastxml}
\alias{blastxml}
\alias{blastxml-package}
\title{blastxml: Extract data from Blast XML files}
\description{
blastxml provides methods for extracting blast resuls in XML files to character martices.
}

