% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{blastxml_info}
\alias{blastxml_info}
\title{Extract blast XML info}
\usage{
blastxml_info(xml_file)
}
\arguments{
\item{xml_file}{character - the path and name of the XML file}
}
\value{
list of
\itemize{
\item{filename}
\item{BlastOutput_program}
\item{BlastOutput_version}
\item{BlastOutput_db}
\item{N_iterations}
}
}
\description{
Extract blast XML info
}
