% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{blastxml_program}
\alias{blastxml_program}
\title{Extract the blast program name from the XML file}
\usage{
blastxml_program(xml_file)
}
\arguments{
\item{xml_file}{character - the path and name of the XML file}
}
\value{
contents of  BlastOutput_program node in xml_file
}
\description{
Extract the blast program name from the XML file
}
