% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump.R
\name{BXD}
\alias{BXD}
\title{A list to identify the Blastfields to retrieve.}
\usage{
BXD(app = "blastn", fields = NULL)
}
\arguments{
\item{app}{character by default the fields are for 'blastn' but this 
is set automatically when the file is opened.}

\item{fields}{character of NULL  if not NULL fields are subset to include 
just these.  The default can be found using \code{get_BXD_names()}}
}
\value{
a named list of fields to return
\itemize{
 \item{app the name of the blast application by default 'blastn'}
 \item{blastn a list of character vectors fields in for Iteration, Hit and Hsp}
}
}
\description{
blastn includes...
 Iteration_iter_num
 Iteration_query_id
 Iteration_query_def
 Iteration_query_len
 Hit_num
 Hit_id
 Hit_def
 Hit_accession
 Hit_len
 Hsp_num
 Hsp_bit_score
 Hsp_score
 Hsp_evalue
 Hsp_query_from
 Hsp_query_to
 Hsp_hit_from
 Hsp_hit_to
 Hsp_query_frame
 Hsp_identity
 Hsp_positive
 Hsp_gaps
 Hsp_align_len
 Hsp_qseq
 Hsp_hseq
 Hsp_midline
}

