% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump.R
\name{get_BXD_names}
\alias{get_BXD_names}
\title{Get the fields for the application}
\usage{
get_BXD_names(bxd = BXD())
}
\arguments{
\item{bxd}{a \code{BXD()} like list}
}
\value{
a character vector
}
\description{
Get the fields for the application
}
