% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump.R
\name{get_example_filename}
\alias{get_example_filename}
\title{Retrieve the path for an example blastn file}
\usage{
get_example_filename(app = "blastn")
}
\arguments{
\item{app}{character, by default 'blastn'}
}
\value{
a filename or empty string
}
\description{
Retrieve the path for an example blastn file
}
