% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump.R
\name{hit_extract}
\alias{hit_extract}
\title{Extract values from an Hit node}
\usage{
hit_extract(x, bxd = BXD())
}
\arguments{
\item{x}{xml2::xml_node for Hit}

\item{bxd}{list of known fields for the blast type, see \code{link{BXD}}}
}
\value{
a character vector of Hit value or NULL
}
\description{
Extract values from an Hit node
}
