% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump.R
\name{hsp_extract}
\alias{hsp_extract}
\title{Extract values from an Hsp node}
\usage{
hsp_extract(x, bxd = BXD())
}
\arguments{
\item{x}{xml2::xml_node for Hsp}

\item{bxd}{list of known fields for the blast type, see \code{link{BXD}}}
}
\value{
a character vector of Hsp value or NULL
}
\description{
Extract values from an Hsp node
}
