% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump.R
\name{iter_extract}
\alias{iter_extract}
\title{Extract values from an Iteration node}
\usage{
iter_extract(x, bxd = BXD())
}
\arguments{
\item{x}{xml2::xml_node for Iteration}

\item{bxd}{list of known fields for the blast type, see \code{link{BXD}}}
}
\value{
a character vector of Iteration value or NULL
}
\description{
Extract values from an Iteration node
}
