% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump.R
\name{set_BXD_names}
\alias{set_BXD_names}
\title{Set the fields to retrieve for the application}
\usage{
set_BXD_names(x, bxd = BXD())
}
\arguments{
\item{x}{a character vector of names or NULL.  If NULL nothing 
is done.}

\item{bxd}{a \code{BXD()} like list}
}
\value{
the possibly updated BXD global
}
\description{
Set the fields to retrieve for the application
}
\examples{
 \dontrun{
     xmlfile <- system.file("extdata", "blastn.xml.gz", package = "blastxml")
     bxd <- BXD(app = 'blastn')
     fields <- get_BXD_names()
     len <- length(fields)
     fields <- fields[-c((len-3):len)]
     bxd <- set_BXD_names(fields, bxd = bxd)
     x <- blastxml_dump(xmlfile, bxd = bxd)
 }  
}

