% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump.R
\name{hsp_extract}
\alias{hsp_extract}
\title{Extract values from an Hsp node}
\usage{
hsp_extract(x, bxd = BXD)
}
\arguments{
\item{x}{xml2::xml_node for Hsp}

\item{nm}{charcater, the Hsp elements to retrieve}
}
\value{
a character vector of Hsp value or NULL
}
\description{
Extract values from an Hsp node
}

